misc/HtmlTestSuite/results/test-0344.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>CSSURLImageValue serialization tests</title>
<link href="https://drafts.css-houdini.org/css-typed-om-1/#urlimagevalue-serialization" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="../resources/testhelper.js"/>
</head><body>
<div id="log"/>
<div id="testUrl" style="background-image: url('/media/1x1-green.png')"/>
<script>
'use strict';

test(() =&gt; {
  const result = document.getElementById('testUrl').attributeStyleMap.get('background-image');
  assert_equals(result.toString(), 'url("/media/1x1-green.png")');
}, 'CSSUrlImageValue serializes correctly');

</script>
</body></html>
